64/code refactoring (#65)
* removed unusued variables time_since_last_ping and methods

* refactored auv_send_data and auv_send_ping threads to separate files

* added constants file

* new file for auv_receive class

* moved globals to constant file and now referencing constants properly

* moved mission command processing out of runloop

* Split base station classes into different files

* add constants file for base station folder

* deleted AUV_Receive from auv.py

* Split up base station files

* fixed lock variable

* create static and threads folders

* reformatting

* update globals file

* rename globals

* split main run loop in auv_receive.py

* added threads folder and static folder

* add case block in run

* remove unusued constants import

* fix threads import

* fix thread import again

* fixed imports for auv_send_data

* old python bad

* fixed lock references in auv_receive and auv_send_data

* fixed references to connected

* remove duplication of statics

* moved locks to constants file and renamed in threads

* shorten run function in base_station_send

* fixed reference to lock in auv_receive

* refactoring

* resolve imports on basestation folder

* add threads.

* constants. in front of PING

* update read from radio

* print in binary

* print message

* dive command

* fix header read

* fix header read
