+++ auv/static/global_vars.py
@@ -1,0 +1,6 @@
+import threading
+# determines if connected to BS
+connected = False
+
+def log(val):
+    print("[AUV]\t" + val)
