+++ base_station/static/globalvars.py
@@ -1,0 +1,3 @@
+# Global variables used by the BaseStation threads
+
+connected = False    # boolean that determines if BS has radio connection with AUV
